requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
exceptiongroup==1.2.1
grpcio==1.62.2
grpcio-health-checking==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
pandas==2.2.2
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.0
weaviate-client==4.5.6
weaviate-demo-datasets==0.4.10
